Update Python files

+++ src/database.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, Text
+from sqlalchemy import create_engine, Column, Integer, String, Text, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
@@ -20,13 +20,11 @@ class Review(Base):
     __tablename__ = "reviews"
 
     id = Column(Integer, primary_key=True, index=True)
-    source = Column(String)  # Ex: site_web, email
-    text_content = Column(Text)  # Le commentaire brut
-    # --- NOUVEAU ---
-    topic = Column(String, nullable=True)  # Pour stocker "Sujet 1" ou "Logistique"
-    # ----------------
-    sentiment = Column(String, nullable=True)  # Pour plus tard (Positif/Négatif)
-    score = Column(Integer, nullable=True)  # Pour plus tard (Note)
+    source = Column(String)
+    text_content = Column(Text)
+    topic = Column(String, nullable=True)
+    sentiment = Column(String, nullable=True) # Stockera "Positif 😃"
+    score = Column(Float, nullable=True)      # <--- CHANGE ICI (Integer -> Float)
 
 # 3. Fonction d'initialisation
 def init_db():
